moved to long_range.py
